fix parallelism build

+++ shared/modeling/src/tensor_parallelism.rs
@@ -40,6 +40,7 @@ impl Default for CommunicatorId {
     }
 }
 
+#[cfg(not(feature = "parallelism"))]
 impl CommunicatorId {
     pub fn new() -> Self {
         unimplemented!()
@@ -197,7 +198,7 @@ impl ColumnParallelLinear {
                 bias,
                 shard: comm.as_ref().map(|comm| Shard {
                     dim: 0,
-                    rank: comm.rank(),
+                    rank: comm.rank() as usize,
                     world_size: comm.size() as usize,
                 }),
                 ..Default::default()
@@ -259,7 +260,7 @@ impl RowParallelLinear {
                 bias,
                 shard: comm.as_ref().map(|comm| Shard {
                     dim: 1,
-                    rank: comm.rank(),
+                    rank: comm.rank() as usize,
                     world_size: comm.size() as usize,
                 }),
                 ..Default::default()
